Move extract_from_device_deployments to utils
