+++ ansible_collections/arista/avd/plugins/action/cv_workflow.py
@@ -42,8 +42,8 @@ try:
         CVWorkspace,
         CVWorkspaceBuildWarningsConfig,
         DeployToCvResult,
-        extract_from_device_deployments,
     )
+    from pyavd._cv.workflows.utils import extract_from_device_deployments
     from pyavd._utils import default, get, strip_empties_from_dict
 
     HAS_PYAVD = True
